mapreduce完成了各职业对各党派的献金情况 修改了总金额的部分bug：金额从string转为int时有小数存在，先转float再转int

+++ aoue_analyzer/src/main/java/com/nd/aoue/analyzer/reducer/SumAmountReducer.java
@@ -15,7 +15,7 @@ public class SumAmountReducer extends Reducer<Text, Text,Text,Text> {
         //统计总捐款金额
         int sumAmount=0;
         for (Text value : values) {
-            int amount=Integer.parseInt(value.toString());
+            int amount=(int)Float.parseFloat(value.toString());
             sumAmount+=amount;
         }
         System.out.println(key + "_"
